requirements.txt
# numpy >= 1.0.0
# absl-py==0.13.0
# pyparsing==2.4.7
# jupyterlab==3.0.14
# scikit-image==0.18.1
# termcolor==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.4
hydra-core==1.1.0
hydra-submitit-launcher==1.1.5
# pandas>=1.3.0
# ipdb==0.13.9
# yapf==0.31.0
# sklearn==0.0
# matplotlib==3.4.2
# opencv-python==4.5.3.56
pillow==7.1.2
pyglet<=1.5.0,>=1.4.0
gym==0.25.2
numpy>=1.10.0
trimesh==3.9.3
pyzmq>=16.0.0
# scikit-image>=0.13.1
# opencv-python>=3.4
# opencv-python==4.3.0.36
opencv-python
#==4.4.0.44
pyyaml>=3.11

dataclasses<=0.6; python_version < '3.7'
carnivalmirror==0.6.2
pipdeptree
duckietown-world-daffy
pipdeptree
typing-extensions
torchvision
tensorboard
wandb
h5py
numba